requirements.txt
typing_extensions>=4.0.0
requests~=2.32.3
beautifulsoup4~=4.12.3
cachetools~=5.3.0
Flask[async]~=3.1.0
Flask-Caching~=2.3.0
Werkzeug~=3.1.3
aiohttp~=3.12.14
aiocache~=0.12.3
Flask-Compress~=1.17
python-dotenv~=1.0.1
gunicorn~=21.2.0
waitress~=3.0.0
SQLAlchemy~=2.0.0
psycopg2-binary~=2.9.0
